✨ feat: Update user schema in swagger.yaml
This commit updates the user schema in the swagger.yaml file by removing the "application/x-www-form-urlencoded" media type and adding an example for the password field encoded in base64 format. It also includes extra tags to validate the password field and make it optional.

The commit adds a custom tag validator for base64 encoding in the `validator.go` file under the `utils` directory. The validator is registered by calling the `registerCustomTagValidators` function, which is also defined in the same file. The actual validation is performed by the `base64Validator` function, which checks if the field's value is a valid base64 encoded string.